Update README.md
The #4 is something I am definitely not happy with, I see the need but the duplication in `@use` is problematic to me. Could the `discover` actually targets the identifier holder instead of the passage ?

I renamed elements for the moment because it seems easier to read. If we end up using something else, I'll be ok with it.